moved abs ingestion out of notebook and into python
Prior to this change the logic for ingesting python was
duplicated in the 2 ingestion note books. Also the
logic for non main and main abs structures was largely
the same yet was written the seperately for the two.

I think initially there wasn't as much certianty as
there is not about their identical format.

This also has the nice benefit of taking logic out
of the untyped notebooks.